Classes SCoPList and SCoP updated.
Now the infrastructure is in place to create an IR from a transformed
polyhedral model in case the model turns out to be valid. Also, now we
can choose one particular SCoP and update only this single SCoP without
breaking the references contained within other SCoPs. Additionally,
putting all the SCoPs in a SCoPList allows us to do statistics on all
SCoPs of a program, such as find the largest SCoP, the number of SCoPs,
the number of statements, and so on.

SCoPList:
* collects all SCoPs of a single program
* refers to the original program and the resulting IR from any transform
* allows us to identify interesting SCoPs
* responsible for overall statistics (effectiveness of polyh transform)

SCoP:
* does not refer to the code, only to the statements
* lists all the statements in a SCoP
* can be transformed