runtime: Made atomic macros take a fixed amount of parameters.
This removes warnings when compiling with -pedantic as we never consumed the remaining arguments in the variadic macros.
The additional parameters available in the GCC sync builtins are not implemented there at the moment and also there is no implementation for this in the Windows or Gemsclaim version of the code.