Changed and updated string literal handling.
The builder now adds leading and trailing double quotes to string literals.
This implied several changes throughout the frontend as well as some tests.
Also the type of the created string literals is a ref now, which caused changes in the core and backend.
As simple string literals are by default created as constant refs now, we also introduced a different way to
  specify string literals in the parser using double leading and trailing quotes in IR strings (""foo"") which
  is necessary in order for one pair of quotes to end up in the string literal eventually.