HAT: Link Haskell runtime system via CMakeLists only
One can specify additional GHC options in the stack.yaml file (per
package). But these flags are used for all targets within a Haskell
package. Specifying the GHC options in the cabal file per target, seems
to be ignored by stack completely.

The insieme-hat binary segfaults if `-lHSrts...` option is supplied to
GHC. Therefore this commit moves the flag setting into the
`CMakeLists.txt` to only add the runtime system for the insieme-analysis
lib.

Additionally the `--force-dirty` flag is passed to ensure it will always
be built with that flag.