scripts: Always call .rb scripts via ruby executable
While I generally prefere to have a shebang inside an executable script
which determines what tool is used to execute the script, this does not
work for Windows. And most of the .rb scripts are also executed under
windows therefore we should always run the ruby executable with the
script as first argument.