all: NumericTypes have a signed type by default now.
This especially affected arrays and all related test strings.
Added suffixes (like "u" or "ll") to the printout with toString as well as to the pretty print.
StringValues of float type don't get the 'f' appended anymore - it will still be printed though.
Implemented support for negative values for enum constants.