core: Fixed bug in IRBuilder lambdaExpr function.
There was a bug which stripped a reference from the passed param type instead of dematerializing the type.
Added function to dematerialize a type.
Added function to check for difference in const and volatile qualifiers only for ref types.
Fixed type_check test to use the new functions.
Added missing test coverage to type_check_test.