progress_estimation: Completely reworked the whole transformation.
It is now based on a bottom up transformation instead of a node mapper.
This way the transformation can be correctly applied in all cases we
encountered until now. Each lambda gets transformed to include the
necessary reporting calls and also gets an annotation attached which
describes how much effort has not been reported inside the lambda.